+++ src/SFManagement.Application/Commands/CreateProjectCommand.cs
@@ -1,3 +1,6 @@
 namespace SFManagement.Application.Commands;
 
-public record CreateProjectCommand(string Name, string? DescriptionMd);
+public record CreateProjectCommand(string Name, string? DescriptionMd)
+{
+    public int CreatedId { get; set; }
+}